Refactored activities into their own packages and slightly renamed a few activities
