+++ app/src/main/AndroidManifest.xml
@@ -15,7 +15,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".LoginActivity"
+            android:name=".activity.login.LoginActivity"
             android:label="@string/app_name"
             android:windowSoftInputMode="adjustResize|stateVisible" >
             <intent-filter>
@@ -25,40 +25,40 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".ProfileActivity"
+            android:name=".activity.playerprofile.ProfileActivity"
             android:label="@string/title_activity_profile_statistics"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" >
         </activity>
         <activity
-            android:name=".CourseActivity"
+            android:name=".activity.course.CoursesActivity"
             android:label="@string/title_activity_courses"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar" >
         </activity>
         <activity
-            android:name=".CourseScorecardActivity"
+            android:name=".activity.coursescorecard.CourseScorecardActivity"
             android:label="@string/title_activity_course_detail"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"
-            android:parentActivityName=".CourseActivity" >
+            android:parentActivityName=".activity.course.CoursesActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.frolfr.frolfrclient.CourseActivity" />
+                android:value=".activity.course.CoursesActivity" />
         </activity>
         <activity
-            android:name=".CourseScorecardDetailActivity"
+            android:name=".activity.scorecard.CourseScorecardDetailActivity"
             android:label="@string/title_activity_scorecard_detail"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"
-            android:parentActivityName=".CourseScorecardActivity" >
+            android:parentActivityName=".activity.coursescorecard.CourseScorecardActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.frolfr.frolfrclient.CourseScorecardActivity" />
+                android:value=".activity.coursescorecard.CourseScorecardActivity" />
         </activity>
-        <activity android:name=".NewGameActivity"
+        <activity android:name=".activity.newgame.NewGameActivity"
             android:label="NewGame"
             android:theme="@style/Theme.AppCompat.Light.NoActionBar"
-            android:parentActivityName=".CourseActivity" >
+            android:parentActivityName=".activity.course.CoursesActivity" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value="com.frolfr.frolfrclient.CourseActivity" />
+                android:value=".activity.course.CoursesActivity" />
         </activity>
     </application>
