+++ app/src/main/java/com/frolfr/frolfrclient/activity/newgame/NewGameFragment.java
@@ -1,6 +1,5 @@
-package com.frolfr.frolfrclient;
+package com.frolfr.frolfrclient.activity.newgame;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
 import android.util.Log;
@@ -8,10 +7,9 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
-import android.widget.ArrayAdapter;
-import android.widget.ListView;
 import android.widget.Spinner;
 
+import com.frolfr.frolfrclient.R;
 import com.frolfr.frolfrclient.entity.Course;
 
 /**
