+++ app/src/main/java/com/frolfr/frolfrclient/activity/playerprofile/ProfileFragment.java
@@ -1,13 +1,13 @@
-package com.frolfr.frolfrclient;
+package com.frolfr.frolfrclient.activity.playerprofile;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.ListView;
 import android.widget.TextView;
+
+import com.frolfr.frolfrclient.R;
 
 /**
  * Created by wowens on 6/14/16.
