+++ app/src/main/java/com/frolfr/frolfrclient/activity/scorecard/CourseScorecardDetailFragment.java
@@ -1,4 +1,4 @@
-package com.frolfr.frolfrclient;
+package com.frolfr.frolfrclient.activity.scorecard;
 
 import android.os.Bundle;
 import android.support.v4.app.Fragment;
@@ -7,8 +7,8 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.ListView;
 
+import com.frolfr.frolfrclient.R;
 import com.frolfr.frolfrclient.entity.RoundScorecard;
-import com.frolfr.frolfrclient.entity.Scorecard;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
