+++ app/src/main/java/com/frolfr/frolfrclient/activity/scorecard/HoleDetailAdapter.java
@@ -1,4 +1,4 @@
-package com.frolfr.frolfrclient;
+package com.frolfr.frolfrclient.activity.scorecard;
 
 import android.app.Activity;
 import android.view.LayoutInflater;
@@ -7,8 +7,8 @@ import android.view.ViewGroup;
 import android.widget.ArrayAdapter;
 import android.widget.TextView;
 
+import com.frolfr.frolfrclient.R;
 import com.frolfr.frolfrclient.entity.HoleDetail;
-import com.frolfr.frolfrclient.entity.Scorecard;
 
 import java.util.List;
