+++ app/src/main/res/layout/fragment_course_scorecard.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.frolfr.frolfrclient.CourseScorecardFragment">
+    tools:context=".activity.coursescorecard.CourseScorecardFragment">
     
     <LinearLayout
         android:id="@+id/courseDetailText"
