+++ app/src/main/res/layout/fragment_course_scorecard_detail.xml
@@ -2,7 +2,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="com.frolfr.frolfrclient.CourseScorecardFragment">
+    tools:context=".activity.coursescorecard.CourseScorecardFragment">
 
 
     <LinearLayout
