+++ app/src/main/res/layout/fragment_courses.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.frolfr.frolfrclient.CoursesFragment">
+    tools:context=".activity.course.CoursesFragment">
 
     <!--
         TODO: accordian-like layout where you expand rows grouped by state?
