+++ app/src/main/res/layout/fragment_profile.xml
@@ -4,7 +4,7 @@
     android:paddingRight="@dimen/activity_horizontal_margin"
     android:paddingTop="@dimen/activity_vertical_margin"
     android:paddingBottom="@dimen/activity_vertical_margin"
-    tools:context="com.frolfr.frolfrclient.ProfileFragment">
+    tools:context=".activity.playerprofile.ProfileFragment">
 
     <TextView android:id="@+id/profile_statistics_text"
         android:text="Replace with stats"
