+++ app/src/main/res/menu/menu_courses.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools" tools:context="com.frolfr.frolfrclient.CourseActivity">
+    xmlns:tools="http://schemas.android.com/tools" tools:context=".activity.course.CoursesActivity">
     <item android:id="@+id/action_newgame" android:title="@string/action_newgame"
         android:orderInCategory="0" app:showAsAction="always"/>
         <!--android:onClick="onNewGameClick"-->
